final_v/requirements.txt
bcrypt==4.1.2
Bootstrap-Flask==2.2.0
click==8.1.7
dnspython==2.6.1
email-validator==2.1.0.post1
Flask==2.2.3
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
greenlet==3.0.3
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
numpy==1.26.4
pandas==2.2.0
pyarrow==15.0.0
python-dateutil==2.8.2
pytz==2024.1
six==1.16.0
SQLAlchemy==2.0.6
typing_extensions==4.9.0
tzdata==2024.1
Werkzeug==2.2.3
WTForms==3.1.2
